requirements.txt
numpy
youtube_dl
click >= 7.0
pyyaml
